requirements.txt
celery==5.2.3
Django==4.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
redis==4.1.1

